migrate all scheduler DAO methods to ScheduleDao
